requirements.txt
openenv-core>=0.2.3
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
datasets>=2.14.0
openai>=1.0.0
requests>=2.31.0
huggingface-hub>=0.19.0
